5) работа с CoordinatorLayout: анимация FloatingActionButton (исчезновение/проявление иконки FAB и изображениея из CollapsingToolbarLayout)
