+++ .idea/misc.xml
@@ -25,6 +25,7 @@
         <entry key="app/src/main/res/layout/bottom_navigation_layout.xml" value="0.20677083333333332" />
         <entry key="app/src/main/res/layout/faragment_dialog_sheet_bottom.xml" value="0.2890625" />
         <entry key="app/src/main/res/layout/fragment_constraint.xml" value="0.2833333333333333" />
+        <entry key="app/src/main/res/layout/fragment_coordinator.xml" value="0.2890625" />
         <entry key="app/src/main/res/layout/fragment_earth.xml" value="0.2890625" />
         <entry key="app/src/main/res/layout/fragment_main.xml" value="0.2890625" />
         <entry key="app/src/main/res/layout/fragment_mars.xml" value="0.2078125" />
