+++ app/src/main/java/com/example/materialdesign/view/layout_maket/coordinator/NestedScrollBehavior.kt
@@ -1,0 +1,27 @@
+package com.example.materialdesign.view.layout_maket.coordinator
+
+import android.content.Context
+import android.util.AttributeSet
+import android.view.View
+import androidx.coordinatorlayout.widget.CoordinatorLayout
+import com.google.android.material.appbar.AppBarLayout
+
+class NestedScrollBehavior(context: Context, attributeSet: AttributeSet) :
+    CoordinatorLayout.Behavior<View>(context, attributeSet) {
+    override fun layoutDependsOn(
+        parent: CoordinatorLayout,
+        child: View,
+        dependency: View
+    ) = dependency is AppBarLayout
+
+
+    override fun onDependentViewChanged(
+        parent: CoordinatorLayout,
+        child: View,
+        dependency: View
+    ): Boolean {
+        val appBar = dependency as AppBarLayout
+        child.y = appBar.height.toFloat() + appBar.y
+        return super.onDependentViewChanged(parent, child, dependency)
+    }
+}
