+++ app/src/main/res/values/dimens.xml
@@ -13,4 +13,6 @@
     <dimen name="layout_height">340dp</dimen>
     <dimen name="drawable_padding_app_compat_text_view">10dp</dimen>
     <dimen name="text_size_app_compat_text_view">20sp</dimen>
+    <dimen name="app_bar_height">220dp</dimen>
+    <dimen name="layout_fab_coordinator">30dp</dimen>
 </resources>
