+++ app/src/main/res/values/strings.xml
@@ -41,4 +41,5 @@
     <string name="coordinator">coordinator</string>
     <string name="motion">motion</string>
 
+
 </resources>
